refactor: 💡 newitems staticpropsfunctions move to newspages lib
create lib/newpages. newitems staticprops functions move to lib/newpages
